requirements.txt
python
python-nmap